+++ packages/backend-core/src/services/aggregate/index.ts
@@ -1,2 +1,3 @@
 export { userAggregateService } from './user.aggregate.service';
 export { billingAggregateService } from './billing.aggregate.service';
+export { anonymousAggregateService } from './anonymous.aggregate.service';